client/requirements.txt
websocket-client
python-dotenv
aiortc
asyncio
